Make KotlinDslScriptsModelSpec a crossVersionTest
